requirements.txt
uvicorn
python-dotenv
fastapi
SQLAlchemy
sqlalchemyseed
bcrypt
pydantic
passlib
starlette
psycopg2